Add files via upload
Still a bit rusty on Git hub so bear with me. 
Here's my current progress on MinConflict, it currently randomly colors the points and counts the number of conflicts each point has. The algorithm still needs to swap conflicting points and look a head to see what swaps are best, but I should be able to finish that tonight.