Add files via upload
It's been updated since the changes to the graph system. Having said that it doesn't quite work. Occasionally it gets held up in a loop on one specific point, and the odds of that happening seem to go up the more points there are. I'm not quite sure what the issue is, but it's almost done.